requirements.txt
Django==3.0.5
mysqlclient=1.4.6
djangorestframework=3.11.0
PyYAML=5.3.1
uritemplate=3.0.1